add method for reversing single byte xor cyphers
the 1 byte keyspace is small enough that we can brute force it. Try all 255 bytes, and for each calculate the log probability of each unigram (using a minimum value of -100 to prevent infinities from making it *weird*). Then sum those log-probs together. The byte resulting in the highest sum string is probably the key