solve challenge 5
introduce a new common function for repeating key XOR and use it to solve challenge 5.

This one was tricky because the formatting in cryptopals makes it look like they want a solution that applies this to two different lines independently. It's a bit of an illusion because the provided solution only looks like it is two lines, forced to break by the width constraints of the DOM element